Refactor to speed up data generation

+++ apps/random_pipeline/random_pipeline_generator.cpp
@@ -143,7 +143,7 @@ public:
     }
 
     void generate() {
-        rng.seed(seed);
+        rng.seed((int)seed);
 
         Func tail = input;
         for (int i = 0; i < stages; i++) {
